Refactor tests
Refactor test and run against diagnostic routes

+++ src/utils/autofac.rs
@@ -36,7 +36,7 @@ pub struct TodayWriter {
     #[shaku(inject)]
     output: Arc<dyn IOutput>,
     today: String,
-    year: usize,
+    year: i32,
 }
 
 impl IDateWriter for TodayWriter {
